requirements.txt
# Core Dependencies
torch>=1.9.0
transformers>=4.11.0
datasets>=1.11.0
numpy>=1.20.0
pandas>=1.3.0
tqdm>=4.62.0
scikit-learn>=0.24.2
pyyaml>=5.4.1
psutil>=5.9.0
matplotlib>=3.4.0
nltk>=3.7
python-dotenv>=0.19.0

# For model serving (optional)
flask>=2.0.1
flask-cors>=3.0.10

# For development (optional)
pytest>=6.2.5
black>=21.9b0
isort>=5.9.3
flake8>=3.9.2